change PostController and delete 1 file in static
