+++ src/main/java/com/challenge/blog/repository/PostRepository.java
@@ -1,6 +1,5 @@
 package com.challenge.blog.repository;
 
-
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
